Crate rusoto_s3

Amazon Simple Storage Service

If you're using the service, you're probably looking for S3Client and S3.

Modules

util
Utility helpers for working with S3

Structs

AbortIncompleteMultipartUpload
Specifies the days since the initiation of an Incomplete Multipart Upload that Lifecycle will wait before permanently removing all parts of the upload.

AbortIncompleteMultipartUploadSerializer

AbortMultipartUploadOutput

AbortMultipartUploadRequest

AccelerateConfiguration

AccelerateConfigurationSerializer

AccessControlPolicy

AccessControlPolicySerializer

AccessControlTranslation
Container for information regarding the access control for replicas.

AccessControlTranslationSerializer

AccountIdSerializer

AllowQuotedRecordDelimiterSerializer

AllowedHeaderSerializer

AllowedHeadersSerializer

AllowedMethodSerializer

AllowedMethodsSerializer

AllowedOriginSerializer

AllowedOriginsSerializer

AnalyticsAndOperator

AnalyticsAndOperatorSerializer

AnalyticsConfiguration

AnalyticsConfigurationSerializer

AnalyticsExportDestination

AnalyticsExportDestinationSerializer

AnalyticsFilter

AnalyticsFilterSerializer

AnalyticsIdSerializer

AnalyticsS3BucketDestination

AnalyticsS3BucketDestinationSerializer

AnalyticsS3ExportFileFormatSerializer

BodySerializer

Bucket

BucketAccelerateStatusSerializer

BucketLifecycleConfiguration

BucketLifecycleConfigurationSerializer

BucketLocationConstraintSerializer

BucketLoggingStatus

BucketLoggingStatusSerializer

BucketLogsPermissionSerializer

BucketNameSerializer

BucketVersioningStatusSerializer

CORSConfiguration

CORSConfigurationSerializer

CORSRule

CORSRuleSerializer

CORSRulesSerializer

CSVInput
Describes how a CSV-formatted input object is formatted.

CSVInputSerializer

CSVOutput
Describes how CSV-formatted results are formatted.

CSVOutputSerializer

CloudFunctionConfiguration

CloudFunctionConfigurationSerializer

CloudFunctionInvocationRoleSerializer

CloudFunctionSerializer

CommentsSerializer

CommonPrefix

CompleteMultipartUploadOutput

CompleteMultipartUploadRequest

CompletedMultipartUpload

CompletedMultipartUploadSerializer

CompletedPart

CompletedPartListSerializer

CompletedPartSerializer

CompressionTypeSerializer

Condition

ConditionSerializer

ContinuationEvent

CopyObjectOutput

CopyObjectRequest

CopyObjectResult

CopyPartResult

CreateBucketConfiguration

CreateBucketConfigurationSerializer

CreateBucketOutput

CreateBucketRequest

CreateMultipartUploadOutput

CreateMultipartUploadRequest

DateSerializer

DaysAfterInitiationSerializer

DaysSerializer

Delete

DeleteBucketAnalyticsConfigurationRequest

DeleteBucketCorsRequest

DeleteBucketEncryptionRequest

DeleteBucketInventoryConfigurationRequest

DeleteBucketLifecycleRequest

DeleteBucketMetricsConfigurationRequest

DeleteBucketPolicyRequest

DeleteBucketReplicationRequest

DeleteBucketRequest

DeleteBucketTaggingRequest

DeleteBucketWebsiteRequest

DeleteMarkerEntry

DeleteObjectOutput

DeleteObjectRequest

DeleteObjectTaggingOutput

DeleteObjectTaggingRequest

DeleteObjectsOutput

DeleteObjectsRequest

DeleteSerializer

DeletedObject

DelimiterSerializer

DescriptionSerializer

Destination
Container for replication destination information.

DestinationSerializer

DisplayNameSerializer

ETagSerializer

EmailAddressSerializer

EnableRequestProgressSerializer

EncodingTypeSerializer

Encryption
Describes the server-side encryption that will be applied to the restore results.

EncryptionConfiguration
Container for information regarding encryption based configuration for replicas.

EncryptionConfigurationSerializer

EncryptionSerializer

EndEvent

ErrorDocument

ErrorDocumentSerializer

EventListSerializer

EventSerializer

ExpirationStatusSerializer

ExpiredObjectDeleteMarkerSerializer

ExposeHeaderSerializer

ExposeHeadersSerializer

ExpressionSerializer

ExpressionTypeSerializer

FetchOwnerSerializer

FieldDelimiterSerializer

FileHeaderInfoSerializer

FilterRule
Container for key value pair that defines the criteria for the filter rule.

FilterRuleListSerializer

FilterRuleNameSerializer

FilterRuleSerializer

FilterRuleValueSerializer

GetBucketAccelerateConfigurationOutput

GetBucketAccelerateConfigurationRequest

GetBucketAclOutput

GetBucketAclRequest

GetBucketAnalyticsConfigurationOutput

GetBucketAnalyticsConfigurationRequest

GetBucketCorsOutput

GetBucketCorsRequest

GetBucketEncryptionOutput

GetBucketEncryptionRequest

GetBucketInventoryConfigurationOutput

GetBucketInventoryConfigurationRequest

GetBucketLifecycleConfigurationOutput

GetBucketLifecycleConfigurationRequest

GetBucketLifecycleOutput

GetBucketLifecycleRequest

GetBucketLocationOutput

GetBucketLocationRequest

GetBucketLoggingOutput

GetBucketLoggingRequest

GetBucketMetricsConfigurationOutput

GetBucketMetricsConfigurationRequest

GetBucketNotificationConfigurationRequest

GetBucketPolicyOutput

GetBucketPolicyRequest

GetBucketReplicationOutput

GetBucketReplicationRequest

GetBucketRequestPaymentOutput

GetBucketRequestPaymentRequest

GetBucketTaggingOutput

GetBucketTaggingRequest

GetBucketVersioningOutput

GetBucketVersioningRequest

GetBucketWebsiteOutput

GetBucketWebsiteRequest

GetObjectAclOutput

GetObjectAclRequest

GetObjectOutput

GetObjectRequest

GetObjectTaggingOutput

GetObjectTaggingRequest

GetObjectTorrentOutput

GetObjectTorrentRequest

GlacierJobParameters

GlacierJobParametersSerializer

Grant

GrantSerializer

Grantee

GranteeSerializer

GrantsSerializer

HeadBucketRequest

HeadObjectOutput

HeadObjectRequest

HostNameSerializer

HttpErrorCodeReturnedEqualsSerializer

HttpRedirectCodeSerializer

IDSerializer

IndexDocument

IndexDocumentSerializer

Initiator

InputSerialization
Describes the serialization format of the object.

InputSerializationSerializer

InventoryConfiguration

InventoryConfigurationSerializer

InventoryDestination

InventoryDestinationSerializer

InventoryEncryption
Contains the type of server-side encryption used to encrypt the inventory results.

InventoryEncryptionSerializer

InventoryFilter

InventoryFilterSerializer

InventoryFormatSerializer

InventoryFrequencySerializer

InventoryIdSerializer

InventoryIncludedObjectVersionsSerializer

InventoryOptionalFieldSerializer

InventoryOptionalFieldsSerializer

InventoryS3BucketDestination

InventoryS3BucketDestinationSerializer

InventorySchedule

InventoryScheduleSerializer

IsEnabledSerializer

JSONInput

JSONInputSerializer

JSONOutput

JSONOutputSerializer

JSONTypeSerializer

KMSContextSerializer

KeyMarkerSerializer

KeyPrefixEqualsSerializer

LambdaFunctionArnSerializer

LambdaFunctionConfiguration
Container for specifying the AWS Lambda notification configuration.

LambdaFunctionConfigurationListSerializer

LambdaFunctionConfigurationSerializer

LifecycleConfiguration

LifecycleConfigurationSerializer

LifecycleExpiration

LifecycleExpirationSerializer

LifecycleRule

LifecycleRuleAndOperator
This is used in a Lifecycle Rule Filter to apply a logical AND to two or more predicates. The Lifecycle Rule will apply to any object matching all of the predicates configured inside the And operator.

LifecycleRuleAndOperatorSerializer

LifecycleRuleFilter
The Filter is used to identify objects that a Lifecycle Rule applies to. A Filter must have exactly one of Prefix, Tag, or And specified.

LifecycleRuleFilterSerializer

LifecycleRuleSerializer

LifecycleRulesSerializer

ListBucketAnalyticsConfigurationsOutput

ListBucketAnalyticsConfigurationsRequest

ListBucketInventoryConfigurationsOutput

ListBucketInventoryConfigurationsRequest

ListBucketMetricsConfigurationsOutput

ListBucketMetricsConfigurationsRequest

ListBucketsOutput

ListMultipartUploadsOutput

ListMultipartUploadsRequest

ListObjectVersionsOutput

ListObjectVersionsRequest

ListObjectsOutput

ListObjectsRequest

ListObjectsV2Output

ListObjectsV2Request

ListPartsOutput

ListPartsRequest

LocationPrefixSerializer

LoggingEnabled
Container for logging information. Presence of this element indicates that logging is enabled. Parameters TargetBucket and TargetPrefix are required in this case.

LoggingEnabledSerializer

MFADeleteSerializer

MarkerSerializer

MaxAgeSecondsSerializer

MaxKeysSerializer

MaxPartsSerializer

MaxUploadsSerializer

MetadataEntry
A metadata key-value pair to store with an object.

MetadataEntrySerializer

MetadataKeySerializer

MetadataValueSerializer

MetricsAndOperator

MetricsAndOperatorSerializer

MetricsConfiguration

MetricsConfigurationSerializer

MetricsFilter

MetricsFilterSerializer

MetricsIdSerializer

MultipartUpload

MultipartUploadIdSerializer

NoncurrentVersionExpiration
Specifies when noncurrent object versions expire. Upon expiration, Amazon S3 permanently deletes the noncurrent object versions. You set this lifecycle configuration action on a bucket that has versioning enabled (or suspended) to request that Amazon S3 delete noncurrent object versions at a specific period in the object's lifetime.

NoncurrentVersionExpirationSerializer

NoncurrentVersionTransition
Container for the transition rule that describes when noncurrent objects transition to the STANDARD_IA, ONEZONE_IA or GLACIER storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to the STANDARD_IA, ONEZONE_IA or GLACIER storage class at a specific period in the object's lifetime.

NoncurrentVersionTransitionListSerializer

NoncurrentVersionTransitionSerializer

NotificationConfiguration
Container for specifying the notification configuration of the bucket. If this element is empty, notifications are turned off on the bucket.

NotificationConfigurationDeprecated

NotificationConfigurationDeprecatedSerializer

NotificationConfigurationFilter
Container for object key name filtering rules. For information about key name filtering, go to Configuring Event Notifications in the Amazon Simple Storage Service Developer Guide.

NotificationConfigurationFilterSerializer

NotificationConfigurationSerializer

NotificationIdSerializer

Object

ObjectCannedACLSerializer

ObjectIdentifier

ObjectIdentifierListSerializer

ObjectIdentifierSerializer

ObjectKeySerializer

ObjectVersion

ObjectVersionIdSerializer

OutputLocation
Describes the location where the restore job's output is stored.

OutputLocationSerializer

OutputSerialization
Describes how results of the Select job are serialized.

OutputSerializationSerializer

Owner

OwnerOverrideSerializer

OwnerSerializer

Part

PartNumberMarkerSerializer

PartNumberSerializer

PayerSerializer

PermissionSerializer

PolicySerializer

PrefixSerializer

Progress

ProgressEvent

ProtocolSerializer

PutBucketAccelerateConfigurationRequest

PutBucketAclRequest

PutBucketAnalyticsConfigurationRequest

PutBucketCorsRequest

PutBucketEncryptionRequest

PutBucketInventoryConfigurationRequest

PutBucketLifecycleConfigurationRequest

PutBucketLifecycleRequest

PutBucketLoggingRequest

PutBucketMetricsConfigurationRequest

PutBucketNotificationConfigurationRequest

PutBucketNotificationRequest

PutBucketPolicyRequest

PutBucketReplicationRequest

PutBucketRequestPaymentRequest

PutBucketTaggingRequest

PutBucketVersioningRequest

PutBucketWebsiteRequest

PutObjectAclOutput

PutObjectAclRequest

PutObjectOutput

PutObjectRequest

PutObjectTaggingOutput

PutObjectTaggingRequest

QueueArnSerializer

QueueConfiguration
Container for specifying an configuration when you want Amazon S3 to publish events to an Amazon Simple Queue Service (Amazon SQS) queue.

QueueConfigurationDeprecated

QueueConfigurationDeprecatedSerializer

QueueConfigurationListSerializer

QueueConfigurationSerializer

QuietSerializer

QuoteCharacterSerializer

QuoteEscapeCharacterSerializer

QuoteFieldsSerializer

RecordDelimiterSerializer

RecordsEvent

Redirect

RedirectAllRequestsTo

RedirectAllRequestsToSerializer

RedirectSerializer

ReplaceKeyPrefixWithSerializer

ReplaceKeyWithSerializer

ReplicaKmsKeyIDSerializer

ReplicationConfiguration
Container for replication rules. You can add as many as 1,000 rules. Total replication configuration size can be up to 2 MB.

ReplicationConfigurationSerializer

ReplicationRule
Container for information about a particular replication rule.

ReplicationRuleSerializer

ReplicationRuleStatusSerializer

ReplicationRulesSerializer

RequestPaymentConfiguration

RequestPaymentConfigurationSerializer

RequestProgress

RequestProgressSerializer

ResponseCacheControlSerializer

ResponseContentDispositionSerializer

ResponseContentEncodingSerializer

ResponseContentLanguageSerializer

ResponseContentTypeSerializer

ResponseExpiresSerializer

RestoreObjectOutput

RestoreObjectRequest

RestoreRequest
Container for restore job parameters.

RestoreRequestSerializer

RestoreRequestTypeSerializer

RoleSerializer

RoutingRule

RoutingRuleSerializer

RoutingRulesSerializer

Rule

RuleSerializer

RulesSerializer

S3Error

S3KeyFilter
Container for object key name prefix and suffix filtering rules.

S3KeyFilterSerializer

S3Location
Describes an S3 location that will receive the results of the restore request.

S3LocationSerializer

S3Client
A client for the Amazon S3 API.

SSEKMS
Specifies the use of SSE-KMS to encrypt delievered Inventory reports.

SSEKMSKeyIdSerializer

SSEKMSSerializer

SSES3
Specifies the use of SSE-S3 to encrypt delievered Inventory reports.

SSES3Serializer

SelectObjectContentEventStream

SelectObjectContentOutput

SelectObjectContentRequest
Request to filter the contents of an Amazon S3 object based on a simple Structured Query Language (SQL) statement. In the request, along with the SQL expression, you must also specify a data serialization format (JSON or CSV) of the object. Amazon S3 uses this to parse object data into records, and returns only records that match the specified SQL expression. You must also specify the data serialization format for the response. For more information, go to S3Select API Documentation.

SelectObjectContentRequestSerializer

SelectParameters
Describes the parameters for Select job types.

SelectParametersSerializer

ServerSideEncryptionByDefault
Describes the default server-side encryption to apply to new objects in the bucket. If Put Object request does not specify any server-side encryption, this default encryption will be applied.

ServerSideEncryptionByDefaultSerializer

ServerSideEncryptionConfiguration
Container for server-side encryption configuration rules. Currently S3 supports one rule only.

ServerSideEncryptionConfigurationSerializer

ServerSideEncryptionRule
Container for information about a particular server-side encryption configuration rule.

ServerSideEncryptionRuleSerializer

ServerSideEncryptionRulesSerializer

ServerSideEncryptionSerializer

SourceSelectionCriteria
Container for filters that define which source objects should be replicated.

SourceSelectionCriteriaSerializer

SseKmsEncryptedObjects
Container for filter information of selection of KMS Encrypted S3 objects.

SseKmsEncryptedObjectsSerializer

SseKmsEncryptedObjectsStatusSerializer

StartAfterSerializer

Stats

StatsEvent

StorageClassAnalysis

StorageClassAnalysisDataExport

StorageClassAnalysisDataExportSerializer

StorageClassAnalysisSchemaVersionSerializer

StorageClassAnalysisSerializer

StorageClassSerializer

SuffixSerializer

Tag

TagSerializer

TagSetSerializer

Tagging

TaggingSerializer

TargetBucketSerializer

TargetGrant

TargetGrantSerializer

TargetGrantsSerializer

TargetPrefixSerializer

TierSerializer

TokenSerializer

TopicArnSerializer

TopicConfiguration
Container for specifying the configuration when you want Amazon S3 to publish events to an Amazon Simple Notification Service (Amazon SNS) topic.

TopicConfigurationDeprecated

TopicConfigurationDeprecatedSerializer

TopicConfigurationListSerializer

TopicConfigurationSerializer

Transition

TransitionListSerializer

TransitionSerializer

TransitionStorageClassSerializer

TypeSerializer

URISerializer

UploadIdMarkerSerializer

UploadPartCopyOutput

UploadPartCopyRequest

UploadPartOutput

UploadPartRequest

UserMetadataSerializer

ValueSerializer

VersionIdMarkerSerializer

VersioningConfiguration

VersioningConfigurationSerializer

WebsiteConfiguration

WebsiteConfigurationSerializer

Enums

AbortMultipartUploadError
Errors returned by AbortMultipartUpload

CompleteMultipartUploadError
Errors returned by CompleteMultipartUpload

CopyObjectError
Errors returned by CopyObject

CreateBucketError
Errors returned by CreateBucket

CreateMultipartUploadError
Errors returned by CreateMultipartUpload

DeleteBucketAnalyticsConfigurationError
Errors returned by DeleteBucketAnalyticsConfiguration

DeleteBucketCorsError
Errors returned by DeleteBucketCors

DeleteBucketEncryptionError
Errors returned by DeleteBucketEncryption

DeleteBucketError
Errors returned by DeleteBucket

DeleteBucketInventoryConfigurationError
Errors returned by DeleteBucketInventoryConfiguration

DeleteBucketLifecycleError
Errors returned by DeleteBucketLifecycle

DeleteBucketMetricsConfigurationError
Errors returned by DeleteBucketMetricsConfiguration

DeleteBucketPolicyError
Errors returned by DeleteBucketPolicy

DeleteBucketReplicationError
Errors returned by DeleteBucketReplication

DeleteBucketTaggingError
Errors returned by DeleteBucketTagging

DeleteBucketWebsiteError
Errors returned by DeleteBucketWebsite

DeleteObjectError
Errors returned by DeleteObject

DeleteObjectTaggingError
Errors returned by DeleteObjectTagging

DeleteObjectsError
Errors returned by DeleteObjects

GetBucketAccelerateConfigurationError
Errors returned by GetBucketAccelerateConfiguration

GetBucketAclError
Errors returned by GetBucketAcl

GetBucketAnalyticsConfigurationError
Errors returned by GetBucketAnalyticsConfiguration

GetBucketCorsError
Errors returned by GetBucketCors

GetBucketEncryptionError
Errors returned by GetBucketEncryption

GetBucketInventoryConfigurationError
Errors returned by GetBucketInventoryConfiguration

GetBucketLifecycleConfigurationError
Errors returned by GetBucketLifecycleConfiguration

GetBucketLifecycleError
Errors returned by GetBucketLifecycle

GetBucketLocationError
Errors returned by GetBucketLocation

GetBucketLoggingError
Errors returned by GetBucketLogging

GetBucketMetricsConfigurationError
Errors returned by GetBucketMetricsConfiguration

GetBucketNotificationConfigurationError
Errors returned by GetBucketNotificationConfiguration

GetBucketNotificationError
Errors returned by GetBucketNotification

GetBucketPolicyError
Errors returned by GetBucketPolicy

GetBucketReplicationError
Errors returned by GetBucketReplication

GetBucketRequestPaymentError
Errors returned by GetBucketRequestPayment

GetBucketTaggingError
Errors returned by GetBucketTagging

GetBucketVersioningError
Errors returned by GetBucketVersioning

GetBucketWebsiteError
Errors returned by GetBucketWebsite

GetObjectAclError
Errors returned by GetObjectAcl

GetObjectError
Errors returned by GetObject

GetObjectTaggingError
Errors returned by GetObjectTagging

GetObjectTorrentError
Errors returned by GetObjectTorrent

HeadBucketError
Errors returned by HeadBucket

HeadObjectError
Errors returned by HeadObject

ListBucketAnalyticsConfigurationsError
Errors returned by ListBucketAnalyticsConfigurations

ListBucketInventoryConfigurationsError
Errors returned by ListBucketInventoryConfigurations

ListBucketMetricsConfigurationsError
Errors returned by ListBucketMetricsConfigurations

ListBucketsError
Errors returned by ListBuckets

ListMultipartUploadsError
Errors returned by ListMultipartUploads

ListObjectVersionsError
Errors returned by ListObjectVersions

ListObjectsError
Errors returned by ListObjects

ListObjectsV2Error
Errors returned by ListObjectsV2

ListPartsError
Errors returned by ListParts

PutBucketAccelerateConfigurationError
Errors returned by PutBucketAccelerateConfiguration

PutBucketAclError
Errors returned by PutBucketAcl

PutBucketAnalyticsConfigurationError
Errors returned by PutBucketAnalyticsConfiguration

PutBucketCorsError
Errors returned by PutBucketCors

PutBucketEncryptionError
Errors returned by PutBucketEncryption

PutBucketInventoryConfigurationError
Errors returned by PutBucketInventoryConfiguration

PutBucketLifecycleConfigurationError
Errors returned by PutBucketLifecycleConfiguration

PutBucketLifecycleError
Errors returned by PutBucketLifecycle

PutBucketLoggingError
Errors returned by PutBucketLogging

PutBucketMetricsConfigurationError
Errors returned by PutBucketMetricsConfiguration

PutBucketNotificationConfigurationError
Errors returned by PutBucketNotificationConfiguration

PutBucketNotificationError
Errors returned by PutBucketNotification

PutBucketPolicyError
Errors returned by PutBucketPolicy

PutBucketReplicationError
Errors returned by PutBucketReplication

PutBucketRequestPaymentError
Errors returned by PutBucketRequestPayment

PutBucketTaggingError
Errors returned by PutBucketTagging

PutBucketVersioningError
Errors returned by PutBucketVersioning

PutBucketWebsiteError
Errors returned by PutBucketWebsite

PutObjectAclError
Errors returned by PutObjectAcl

PutObjectError
Errors returned by PutObject

PutObjectTaggingError
Errors returned by PutObjectTagging

RestoreObjectError
Errors returned by RestoreObject

SelectObjectContentError
Errors returned by SelectObjectContent

UploadPartCopyError
Errors returned by UploadPartCopy

UploadPartError
Errors returned by UploadPart

Traits

S3
Trait representing the capabilities of the Amazon S3 API. Amazon S3 clients implement this trait.

Type Definitions

StreamingBody