Struct rusoto_s3::GetObjectOutput

pub struct GetObjectOutput {
    pub accept_ranges: Option<String>,
    pub body: Option<StreamingBody>,
    pub bucket_key_enabled: Option<bool>,
    pub cache_control: Option<String>,
    pub content_disposition: Option<String>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_length: Option<i64>,
    pub content_range: Option<String>,
    pub content_type: Option<String>,
    pub delete_marker: Option<bool>,
    pub e_tag: Option<String>,
    pub expiration: Option<String>,
    pub expires: Option<String>,
    pub last_modified: Option<String>,
    pub metadata: Option<HashMap<String, String>>,
    pub missing_meta: Option<i64>,
    pub object_lock_legal_hold_status: Option<String>,
    pub object_lock_mode: Option<String>,
    pub object_lock_retain_until_date: Option<String>,
    pub parts_count: Option<i64>,
    pub replication_status: Option<String>,
    pub request_charged: Option<String>,
    pub restore: Option<String>,
    pub sse_customer_algorithm: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub ssekms_key_id: Option<String>,
    pub server_side_encryption: Option<String>,
    pub storage_class: Option<String>,
    pub tag_count: Option<i64>,
    pub version_id: Option<String>,
    pub website_redirect_location: Option<String>,
}

Fields
accept_ranges: Option<String>

Indicates that a range of bytes was specified.

body: Option<StreamingBody>

Object data.

bucket_key_enabled: Option<bool>

Indicates whether the object uses an S3 Bucket Key for server-side encryption with AWS KMS (SSE-KMS).

cache_control: Option<String>

Specifies caching behavior along the request/reply chain.

content_disposition: Option<String>

Specifies presentational information for the object.

content_encoding: Option<String>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

content_language: Option<String>

The language the content is in.

content_length: Option<i64>

Size of the body in bytes.

content_range: Option<String>

The portion of the object returned in the response.

content_type: Option<String>

A standard MIME type describing the format of the object data.

delete_marker: Option<bool>

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

e_tag: Option<String>

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

expiration: Option<String>

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

expires: Option<String>

The date and time at which the object is no longer cacheable.

last_modified: Option<String>

Last modified date of the object

metadata: Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

missing_meta: Option<i64>

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

object_lock_legal_hold_status: Option<String>

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

object_lock_mode: Option<String>

The Object Lock mode currently in place for this object.

object_lock_retain_until_date: Option<String>

The date and time when this object's Object Lock will expire.

parts_count: Option<i64>

The count of parts this object has.

replication_status: Option<String>

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

request_charged: Option<String>restore: Option<String>

Provides information about object restoration operation and expiration time of the restored object copy.

sse_customer_algorithm: Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

sse_customer_key_md5: Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

ssekms_key_id: Option<String>

If present, specifies the ID of the AWS Key Management Service (AWS KMS) symmetric customer managed customer master key (CMK) that was used for the object.

server_side_encryption: Option<String>

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

storage_class: Option<String>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

tag_count: Option<i64>

The number of tags, if any, on the object.

version_id: Option<String>

Version of the object.

website_redirect_location: Option<String>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.