Struct rusoto_s3::GetObjectRequest

pub struct GetObjectRequest {
    pub bucket: String,
    pub if_match: Option<String>,
    pub if_modified_since: Option<String>,
    pub if_none_match: Option<String>,
    pub if_unmodified_since: Option<String>,
    pub key: String,
    pub part_number: Option<i64>,
    pub range: Option<String>,
    pub request_payer: Option<String>,
    pub response_cache_control: Option<String>,
    pub response_content_disposition: Option<String>,
    pub response_content_encoding: Option<String>,
    pub response_content_language: Option<String>,
    pub response_content_type: Option<String>,
    pub response_expires: Option<String>,
    pub sse_customer_algorithm: Option<String>,
    pub sse_customer_key: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub version_id: Option<String>,
}

Fields
bucket: String if_match: Option<String>

Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

if_modified_since: Option<String>

Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

if_none_match: Option<String>

Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

if_unmodified_since: Option<String>

Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

key: String part_number: Option<i64>

Part number of the object being read. This is a positive integer between 1 and 10,000. Effectively performs a 'ranged' GET request for the part specified. Useful for downloading just a part of an object.

range: Option<String>

Downloads the specified range bytes of an object. For more information about the HTTP Range header, go to http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

request_payer: Option<String> response_cache_control: Option<String>

Sets the Cache-Control header of the response.

response_content_disposition: Option<String>

Sets the Content-Disposition header of the response

response_content_encoding: Option<String>

Sets the Content-Encoding header of the response.

response_content_language: Option<String>

Sets the Content-Language header of the response.

response_content_type: Option<String>

Sets the Content-Type header of the response.

response_expires: Option<String>

Sets the Expires header of the response.

sse_customer_algorithm: Option<String>

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

sse_customer_key: Option<String>

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

sse_customer_key_md5: Option<String>

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

version_id: Option<String>

VersionId used to reference a specific version of the object.