Struct rusoto_s3::HeadObjectOutput

pub struct HeadObjectOutput {
    pub accept_ranges: Option<String>,
    pub archive_status: Option<String>,
    pub bucket_key_enabled: Option<bool>,
    pub cache_control: Option<String>,
    pub content_disposition: Option<String>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_length: Option<i64>,
    pub content_type: Option<String>,
    pub delete_marker: Option<bool>,
    pub e_tag: Option<String>,
    pub expiration: Option<String>,
    pub expires: Option<String>,
    pub last_modified: Option<String>,
    pub metadata: Option<HashMap<String, String>>,
    pub missing_meta: Option<i64>,
    pub object_lock_legal_hold_status: Option<String>,
    pub object_lock_mode: Option<String>,
    pub object_lock_retain_until_date: Option<String>,
    pub parts_count: Option<i64>,
    pub replication_status: Option<String>,
    pub request_charged: Option<String>,
    pub restore: Option<String>,
    pub sse_customer_algorithm: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub ssekms_key_id: Option<String>,
    pub server_side_encryption: Option<String>,
    pub storage_class: Option<String>,
    pub version_id: Option<String>,
    pub website_redirect_location: Option<String>,
}

Fields
accept_ranges: Option<String>

Indicates that a range of bytes was specified.

archive_status: Option<String>

The archive state of the head object.

bucket_key_enabled: Option<bool>

Indicates whether the object uses an S3 Bucket Key for server-side encryption with AWS KMS (SSE-KMS).

cache_control: Option<String>

Specifies caching behavior along the request/reply chain.

content_disposition: Option<String>

Specifies presentational information for the object.

content_encoding: Option<String>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

content_language: Option<String>

The language the content is in.

content_length: Option<i64>

Size of the body in bytes.

content_type: Option<String>

A standard MIME type describing the format of the object data.

delete_marker: Option<bool>

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

e_tag: Option<String>

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

expiration: Option<String>

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

expires: Option<String>

The date and time at which the object is no longer cacheable.

last_modified: Option<String>

Last modified date of the object

metadata: Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

missing_meta: Option<i64>

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

object_lock_legal_hold_status: Option<String>

Specifies whether a legal hold is in effect for this object. This header is only returned if the requester has the s3:GetObjectLegalHold permission. This header is not returned if the specified version of this object has never had a legal hold applied. For more information about S3 Object Lock, see Object Lock.

object_lock_mode: Option<String>

The Object Lock mode, if any, that's in effect for this object. This header is only returned if the requester has the s3:GetObjectRetention permission. For more information about S3 Object Lock, see Object Lock.

object_lock_retain_until_date: Option<String>

The date and time when the Object Lock retention period expires. This header is only returned if the requester has the s3:GetObjectRetention permission.

parts_count: Option<i64>

The count of parts this object has.

replication_status: Option<String>

Amazon S3 can return this header if your request involves a bucket that is either a source or a destination in a replication rule.

In replication, you have a source bucket on which you configure replication and destination bucket or buckets where Amazon S3 stores object replicas. When you request an object (GetObject) or object metadata (HeadObject) from these buckets, Amazon S3 will return the x-amz-replication-status header in the response as follows:

If requesting an object from the source bucket — Amazon S3 will return the x-amz-replication-status header if the object in your request is eligible for replication.

For example, suppose that in your replication configuration, you specify object prefix TaxDocs requesting Amazon S3 to replicate objects with key prefix TaxDocs. Any objects you upload with this key name prefix, for example TaxDocs/document1.pdf, are eligible for replication. For any object request with this key name prefix, Amazon S3 will return the x-amz-replication-status header with value PENDING, COMPLETED or FAILED indicating object replication status.

If requesting an object from a destination bucket — Amazon S3 will return the x-amz-replication-status header with value REPLICA if the object in your request is a replica that Amazon S3 created and there is no replica modification replication in progress.

When replicating objects to multiple destination buckets the x-amz-replication-status header acts differently. The header of the source object will only return a value of COMPLETED when replication is successful to all destinations. The header will remain at value PENDING until replication has completed for all destinations. If one or more destinations fails replication the header will return FAILED.

For more information, see Replication.

request_charged: Option<String>restore: Option<String>

If the object is an archived object (an object whose storage class is GLACIER), the response includes this header if either the archive restoration is in progress (see RestoreObject or an archive copy is already restored.

If an archive copy is already restored, the header value indicates when Amazon S3 is scheduled to delete the object copy. For example:

x-amz-restore: ongoing-request="false", expiry-date="Fri, 23 Dec 2012 00:00:00 GMT"

If the object restoration is in progress, the header returns the value ongoing-request="true".

For more information about archiving objects, see Transitioning Objects: General Considerations.

sse_customer_algorithm: Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

sse_customer_key_md5: Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

ssekms_key_id: Option<String>

If present, specifies the ID of the AWS Key Management Service (AWS KMS) symmetric customer managed customer master key (CMK) that was used for the object.

server_side_encryption: Option<String>

If the object is stored using server-side encryption either with an AWS KMS customer master key (CMK) or an Amazon S3-managed encryption key, the response includes this header with the value of the server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

storage_class: Option<String>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

version_id: Option<String>

Version of the object.

website_redirect_location: Option<String>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.