Struct rusoto_s3::ListObjectsV2Output

pub struct ListObjectsV2Output {
    pub common_prefixes: Option<Vec<CommonPrefix>>,
    pub contents: Option<Vec<Object>>,
    pub continuation_token: Option<String>,
    pub delimiter: Option<String>,
    pub encoding_type: Option<String>,
    pub is_truncated: Option<bool>,
    pub key_count: Option<i64>,
    pub max_keys: Option<i64>,
    pub name: Option<String>,
    pub next_continuation_token: Option<String>,
    pub prefix: Option<String>,
    pub start_after: Option<String>,
}

Fields
common_prefixes: Option<Vec<CommonPrefix>>

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by delimiter

contents: Option<Vec<Object>>

Metadata about each object returned.

continuation_token: Option<String>

ContinuationToken indicates Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key

delimiter: Option<String>

A delimiter is a character you use to group keys.

encoding_type: Option<String>

Encoding type used by Amazon S3 to encode object keys in the response.

is_truncated: Option<bool>

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

key_count: Option<i64>

KeyCount is the number of keys returned with this request. KeyCount will always be less than equals to MaxKeys field. Say you ask for 50 keys, your result will include less than equals 50 keys

max_keys: Option<i64>

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

name: Option<String>

Name of the bucket to list.

next_continuation_token: Option<String>

NextContinuationToken is sent when isTruncated is true which means there are more keys in the bucket that can be listed. The next list requests to Amazon S3 can be continued with this NextContinuationToken. NextContinuationToken is obfuscated and is not a real key

prefix: Option<String>

Limits the response to keys that begin with the specified prefix.

start_after: Option<String>

StartAfter is where you want Amazon S3 to start listing from. Amazon S3 starts listing after this specified key. StartAfter can be any key in the bucket