Struct rusoto_s3::PutObjectAclRequest

pub struct PutObjectAclRequest {
    pub acl: Option<String>,
    pub access_control_policy: Option<AccessControlPolicy>,
    pub bucket: String,
    pub content_md5: Option<String>,
    pub grant_full_control: Option<String>,
    pub grant_read: Option<String>,
    pub grant_read_acp: Option<String>,
    pub grant_write: Option<String>,
    pub grant_write_acp: Option<String>,
    pub key: String,
    pub request_payer: Option<String>,
    pub version_id: Option<String>,
}

Fields
acl: Option<String>

The canned ACL to apply to the object.

access_control_policy: Option<AccessControlPolicy> bucket: String content_md5: Option<String> grant_full_control: Option<String>

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

grant_read: Option<String>

Allows grantee to list the objects in the bucket.

grant_read_acp: Option<String>

Allows grantee to read the bucket ACL.

grant_write: Option<String>

Allows grantee to create, overwrite, and delete any object in the bucket.

grant_write_acp: Option<String>

Allows grantee to write the ACL for the applicable bucket.

key: String request_payer: Option<String> version_id: Option<String>

VersionId used to reference a specific version of the object.