Struct rusoto_s3::UploadPartCopyRequest

pub struct UploadPartCopyRequest {
    pub bucket: String,
    pub copy_source: String,
    pub copy_source_if_match: Option<String>,
    pub copy_source_if_modified_since: Option<String>,
    pub copy_source_if_none_match: Option<String>,
    pub copy_source_if_unmodified_since: Option<String>,
    pub copy_source_range: Option<String>,
    pub copy_source_sse_customer_algorithm: Option<String>,
    pub copy_source_sse_customer_key: Option<String>,
    pub copy_source_sse_customer_key_md5: Option<String>,
    pub key: String,
    pub part_number: i64,
    pub request_payer: Option<String>,
    pub sse_customer_algorithm: Option<String>,
    pub sse_customer_key: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub upload_id: String,
}

Fields
bucket: String copy_source: String

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

copy_source_if_match: Option<String>

Copies the object if its entity tag (ETag) matches the specified tag.

copy_source_if_modified_since: Option<String>

Copies the object if it has been modified since the specified time.

copy_source_if_none_match: Option<String>

Copies the object if its entity tag (ETag) is different than the specified ETag.

copy_source_if_unmodified_since: Option<String>

Copies the object if it hasn't been modified since the specified time.

copy_source_range: Option<String>

The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first ten bytes of the source. You can copy a range only if the source object is greater than 5 GB.

copy_source_sse_customer_algorithm: Option<String>

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

copy_source_sse_customer_key: Option<String>

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

copy_source_sse_customer_key_md5: Option<String>

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

key: String part_number: i64

Part number of part being copied. This is a positive integer between 1 and 10,000.

request_payer: Option<String> sse_customer_algorithm: Option<String>

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

sse_customer_key: Option<String>

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

sse_customer_key_md5: Option<String>

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upload_id: String

Upload ID identifying the multipart upload whose part is being copied.