Trait rusoto_s3::S3

pub trait S3 {
    fn abort_multipart_upload(
        &self, 
        input: &AbortMultipartUploadRequest
    ) -> Result<AbortMultipartUploadOutput, AbortMultipartUploadError>;
    fn complete_multipart_upload(
        &self, 
        input: &CompleteMultipartUploadRequest
    ) -> Result<CompleteMultipartUploadOutput, CompleteMultipartUploadError>;
    fn copy_object(
        &self, 
        input: &CopyObjectRequest
    ) -> Result<CopyObjectOutput, CopyObjectError>;
    fn create_bucket(
        &self, 
        input: &CreateBucketRequest
    ) -> Result<CreateBucketOutput, CreateBucketError>;
    fn create_multipart_upload(
        &self, 
        input: &CreateMultipartUploadRequest
    ) -> Result<CreateMultipartUploadOutput, CreateMultipartUploadError>;
    fn delete_bucket(
        &self, 
        input: &DeleteBucketRequest
    ) -> Result<(), DeleteBucketError>;
    fn delete_bucket_analytics_configuration(
        &self, 
        input: &DeleteBucketAnalyticsConfigurationRequest
    ) -> Result<(), DeleteBucketAnalyticsConfigurationError>;
    fn delete_bucket_cors(
        &self, 
        input: &DeleteBucketCorsRequest
    ) -> Result<(), DeleteBucketCorsError>;
    fn delete_bucket_inventory_configuration(
        &self, 
        input: &DeleteBucketInventoryConfigurationRequest
    ) -> Result<(), DeleteBucketInventoryConfigurationError>;
    fn delete_bucket_lifecycle(
        &self, 
        input: &DeleteBucketLifecycleRequest
    ) -> Result<(), DeleteBucketLifecycleError>;
    fn delete_bucket_metrics_configuration(
        &self, 
        input: &DeleteBucketMetricsConfigurationRequest
    ) -> Result<(), DeleteBucketMetricsConfigurationError>;
    fn delete_bucket_policy(
        &self, 
        input: &DeleteBucketPolicyRequest
    ) -> Result<(), DeleteBucketPolicyError>;
    fn delete_bucket_replication(
        &self, 
        input: &DeleteBucketReplicationRequest
    ) -> Result<(), DeleteBucketReplicationError>;
    fn delete_bucket_tagging(
        &self, 
        input: &DeleteBucketTaggingRequest
    ) -> Result<(), DeleteBucketTaggingError>;
    fn delete_bucket_website(
        &self, 
        input: &DeleteBucketWebsiteRequest
    ) -> Result<(), DeleteBucketWebsiteError>;
    fn delete_object(
        &self, 
        input: &DeleteObjectRequest
    ) -> Result<DeleteObjectOutput, DeleteObjectError>;
    fn delete_object_tagging(
        &self, 
        input: &DeleteObjectTaggingRequest
    ) -> Result<DeleteObjectTaggingOutput, DeleteObjectTaggingError>;
    fn delete_objects(
        &self, 
        input: &DeleteObjectsRequest
    ) -> Result<DeleteObjectsOutput, DeleteObjectsError>;
    fn get_bucket_accelerate_configuration(
        &self, 
        input: &GetBucketAccelerateConfigurationRequest
    ) -> Result<GetBucketAccelerateConfigurationOutput, GetBucketAccelerateConfigurationError>;
    fn get_bucket_acl(
        &self, 
        input: &GetBucketAclRequest
    ) -> Result<GetBucketAclOutput, GetBucketAclError>;
    fn get_bucket_analytics_configuration(
        &self, 
        input: &GetBucketAnalyticsConfigurationRequest
    ) -> Result<GetBucketAnalyticsConfigurationOutput, GetBucketAnalyticsConfigurationError>;
    fn get_bucket_cors(
        &self, 
        input: &GetBucketCorsRequest
    ) -> Result<GetBucketCorsOutput, GetBucketCorsError>;
    fn get_bucket_inventory_configuration(
        &self, 
        input: &GetBucketInventoryConfigurationRequest
    ) -> Result<GetBucketInventoryConfigurationOutput, GetBucketInventoryConfigurationError>;
    fn get_bucket_lifecycle(
        &self, 
        input: &GetBucketLifecycleRequest
    ) -> Result<GetBucketLifecycleOutput, GetBucketLifecycleError>;
    fn get_bucket_lifecycle_configuration(
        &self, 
        input: &GetBucketLifecycleConfigurationRequest
    ) -> Result<GetBucketLifecycleConfigurationOutput, GetBucketLifecycleConfigurationError>;
    fn get_bucket_location(
        &self, 
        input: &GetBucketLocationRequest
    ) -> Result<GetBucketLocationOutput, GetBucketLocationError>;
    fn get_bucket_logging(
        &self, 
        input: &GetBucketLoggingRequest
    ) -> Result<GetBucketLoggingOutput, GetBucketLoggingError>;
    fn get_bucket_metrics_configuration(
        &self, 
        input: &GetBucketMetricsConfigurationRequest
    ) -> Result<GetBucketMetricsConfigurationOutput, GetBucketMetricsConfigurationError>;
    fn get_bucket_notification(
        &self, 
        input: &GetBucketNotificationConfigurationRequest
    ) -> Result<NotificationConfigurationDeprecated, GetBucketNotificationError>;
    fn get_bucket_notification_configuration(
        &self, 
        input: &GetBucketNotificationConfigurationRequest
    ) -> Result<NotificationConfiguration, GetBucketNotificationConfigurationError>;
    fn get_bucket_policy(
        &self, 
        input: &GetBucketPolicyRequest
    ) -> Result<GetBucketPolicyOutput, GetBucketPolicyError>;
    fn get_bucket_replication(
        &self, 
        input: &GetBucketReplicationRequest
    ) -> Result<GetBucketReplicationOutput, GetBucketReplicationError>;
    fn get_bucket_request_payment(
        &self, 
        input: &GetBucketRequestPaymentRequest
    ) -> Result<GetBucketRequestPaymentOutput, GetBucketRequestPaymentError>;
    fn get_bucket_tagging(
        &self, 
        input: &GetBucketTaggingRequest
    ) -> Result<GetBucketTaggingOutput, GetBucketTaggingError>;
    fn get_bucket_versioning(
        &self, 
        input: &GetBucketVersioningRequest
    ) -> Result<GetBucketVersioningOutput, GetBucketVersioningError>;
    fn get_bucket_website(
        &self, 
        input: &GetBucketWebsiteRequest
    ) -> Result<GetBucketWebsiteOutput, GetBucketWebsiteError>;
    fn get_object(
        &self, 
        input: &GetObjectRequest
    ) -> Result<GetObjectOutput, GetObjectError>;
    fn get_object_acl(
        &self, 
        input: &GetObjectAclRequest
    ) -> Result<GetObjectAclOutput, GetObjectAclError>;
    fn get_object_tagging(
        &self, 
        input: &GetObjectTaggingRequest
    ) -> Result<GetObjectTaggingOutput, GetObjectTaggingError>;
    fn get_object_torrent(
        &self, 
        input: &GetObjectTorrentRequest
    ) -> Result<GetObjectTorrentOutput, GetObjectTorrentError>;
    fn head_bucket(
        &self, 
        input: &HeadBucketRequest
    ) -> Result<(), HeadBucketError>;
    fn head_object(
        &self, 
        input: &HeadObjectRequest
    ) -> Result<HeadObjectOutput, HeadObjectError>;
    fn list_bucket_analytics_configurations(
        &self, 
        input: &ListBucketAnalyticsConfigurationsRequest
    ) -> Result<ListBucketAnalyticsConfigurationsOutput, ListBucketAnalyticsConfigurationsError>;
    fn list_bucket_inventory_configurations(
        &self, 
        input: &ListBucketInventoryConfigurationsRequest
    ) -> Result<ListBucketInventoryConfigurationsOutput, ListBucketInventoryConfigurationsError>;
    fn list_bucket_metrics_configurations(
        &self, 
        input: &ListBucketMetricsConfigurationsRequest
    ) -> Result<ListBucketMetricsConfigurationsOutput, ListBucketMetricsConfigurationsError>;
    fn list_buckets(&self) -> Result<ListBucketsOutput, ListBucketsError>;
    fn list_multipart_uploads(
        &self, 
        input: &ListMultipartUploadsRequest
    ) -> Result<ListMultipartUploadsOutput, ListMultipartUploadsError>;
    fn list_object_versions(
        &self, 
        input: &ListObjectVersionsRequest
    ) -> Result<ListObjectVersionsOutput, ListObjectVersionsError>;
    fn list_objects(
        &self, 
        input: &ListObjectsRequest
    ) -> Result<ListObjectsOutput, ListObjectsError>;
    fn list_objects_v2(
        &self, 
        input: &ListObjectsV2Request
    ) -> Result<ListObjectsV2Output, ListObjectsV2Error>;
    fn list_parts(
        &self, 
        input: &ListPartsRequest
    ) -> Result<ListPartsOutput, ListPartsError>;
    fn put_bucket_accelerate_configuration(
        &self, 
        input: &PutBucketAccelerateConfigurationRequest
    ) -> Result<(), PutBucketAccelerateConfigurationError>;
    fn put_bucket_acl(
        &self, 
        input: &PutBucketAclRequest
    ) -> Result<(), PutBucketAclError>;
    fn put_bucket_analytics_configuration(
        &self, 
        input: &PutBucketAnalyticsConfigurationRequest
    ) -> Result<(), PutBucketAnalyticsConfigurationError>;
    fn put_bucket_cors(
        &self, 
        input: &PutBucketCorsRequest
    ) -> Result<(), PutBucketCorsError>;
    fn put_bucket_inventory_configuration(
        &self, 
        input: &PutBucketInventoryConfigurationRequest
    ) -> Result<(), PutBucketInventoryConfigurationError>;
    fn put_bucket_lifecycle(
        &self, 
        input: &PutBucketLifecycleRequest
    ) -> Result<(), PutBucketLifecycleError>;
    fn put_bucket_lifecycle_configuration(
        &self, 
        input: &PutBucketLifecycleConfigurationRequest
    ) -> Result<(), PutBucketLifecycleConfigurationError>;
    fn put_bucket_logging(
        &self, 
        input: &PutBucketLoggingRequest
    ) -> Result<(), PutBucketLoggingError>;
    fn put_bucket_metrics_configuration(
        &self, 
        input: &PutBucketMetricsConfigurationRequest
    ) -> Result<(), PutBucketMetricsConfigurationError>;
    fn put_bucket_notification(
        &self, 
        input: &PutBucketNotificationRequest
    ) -> Result<(), PutBucketNotificationError>;
    fn put_bucket_notification_configuration(
        &self, 
        input: &PutBucketNotificationConfigurationRequest
    ) -> Result<(), PutBucketNotificationConfigurationError>;
    fn put_bucket_policy(
        &self, 
        input: &PutBucketPolicyRequest
    ) -> Result<(), PutBucketPolicyError>;
    fn put_bucket_replication(
        &self, 
        input: &PutBucketReplicationRequest
    ) -> Result<(), PutBucketReplicationError>;
    fn put_bucket_request_payment(
        &self, 
        input: &PutBucketRequestPaymentRequest
    ) -> Result<(), PutBucketRequestPaymentError>;
    fn put_bucket_tagging(
        &self, 
        input: &PutBucketTaggingRequest
    ) -> Result<(), PutBucketTaggingError>;
    fn put_bucket_versioning(
        &self, 
        input: &PutBucketVersioningRequest
    ) -> Result<(), PutBucketVersioningError>;
    fn put_bucket_website(
        &self, 
        input: &PutBucketWebsiteRequest
    ) -> Result<(), PutBucketWebsiteError>;
    fn put_object(
        &self, 
        input: &PutObjectRequest
    ) -> Result<PutObjectOutput, PutObjectError>;
    fn put_object_acl(
        &self, 
        input: &PutObjectAclRequest
    ) -> Result<PutObjectAclOutput, PutObjectAclError>;
    fn put_object_tagging(
        &self, 
        input: &PutObjectTaggingRequest
    ) -> Result<PutObjectTaggingOutput, PutObjectTaggingError>;
    fn restore_object(
        &self, 
        input: &RestoreObjectRequest
    ) -> Result<RestoreObjectOutput, RestoreObjectError>;
    fn upload_part(
        &self, 
        input: &UploadPartRequest
    ) -> Result<UploadPartOutput, UploadPartError>;
    fn upload_part_copy(
        &self, 
        input: &UploadPartCopyRequest
    ) -> Result<UploadPartCopyOutput, UploadPartCopyError>;
}

Trait representing the capabilities of the Amazon S3 API. Amazon S3 clients implement this trait.

Required Methods

fn abort_multipart_upload(
&self,
input: &AbortMultipartUploadRequest
) -> Result<AbortMultipartUploadOutput, AbortMultipartUploadError>

Aborts a multipart upload.

To verify that all parts have been removed, so you don't get charged for the part storage, you should call the List Parts operation and ensure the parts list is empty.

fn complete_multipart_upload(
&self,
input: &CompleteMultipartUploadRequest
) -> Result<CompleteMultipartUploadOutput, CompleteMultipartUploadError>

Completes a multipart upload by assembling previously uploaded parts.

fn copy_object(
&self,
input: &CopyObjectRequest
) -> Result<CopyObjectOutput, CopyObjectError>

Creates a copy of an object that is already stored in Amazon S3.

fn create_bucket(
&self,
input: &CreateBucketRequest
) -> Result<CreateBucketOutput, CreateBucketError>

Creates a new bucket.

fn create_multipart_upload(
&self,
input: &CreateMultipartUploadRequest
) -> Result<CreateMultipartUploadOutput, CreateMultipartUploadError>

Initiates a multipart upload and returns an upload ID.

Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

fn delete_bucket(
&self,
input: &DeleteBucketRequest
) -> Result<(), DeleteBucketError>

Deletes the bucket. All objects (including all object versions and Delete Markers) in the bucket must be deleted before the bucket itself can be deleted.

fn delete_bucket_analytics_configuration(
&self,
input: &DeleteBucketAnalyticsConfigurationRequest
) -> Result<(), DeleteBucketAnalyticsConfigurationError>

Deletes an analytics configuration for the bucket (specified by the analytics configuration ID).

fn delete_bucket_cors(
&self,
input: &DeleteBucketCorsRequest
) -> Result<(), DeleteBucketCorsError>

Deletes the cors configuration information set for the bucket.

fn delete_bucket_inventory_configuration(
&self,
input: &DeleteBucketInventoryConfigurationRequest
) -> Result<(), DeleteBucketInventoryConfigurationError>

Deletes an inventory configuration (identified by the inventory ID) from the bucket.

fn delete_bucket_lifecycle(
&self,
input: &DeleteBucketLifecycleRequest
) -> Result<(), DeleteBucketLifecycleError>

Deletes the lifecycle configuration from the bucket.

fn delete_bucket_metrics_configuration(
&self,
input: &DeleteBucketMetricsConfigurationRequest
) -> Result<(), DeleteBucketMetricsConfigurationError>

Deletes a metrics configuration (specified by the metrics configuration ID) from the bucket.

fn delete_bucket_policy(
&self,
input: &DeleteBucketPolicyRequest
) -> Result<(), DeleteBucketPolicyError>

Deletes the policy from the bucket.

fn delete_bucket_replication(
&self,
input: &DeleteBucketReplicationRequest
) -> Result<(), DeleteBucketReplicationError>

Deletes the replication configuration from the bucket.

fn delete_bucket_tagging(
&self,
input: &DeleteBucketTaggingRequest
) -> Result<(), DeleteBucketTaggingError>

Deletes the tags from the bucket.

fn delete_bucket_website(
&self,
input: &DeleteBucketWebsiteRequest
) -> Result<(), DeleteBucketWebsiteError>

This operation removes the website configuration from the bucket.

fn delete_object(
&self,
input: &DeleteObjectRequest
) -> Result<DeleteObjectOutput, DeleteObjectError>

Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object. If there isn't a null version, Amazon S3 does not remove any objects.

fn delete_object_tagging(
&self,
input: &DeleteObjectTaggingRequest
) -> Result<DeleteObjectTaggingOutput, DeleteObjectTaggingError>

Removes the tag-set from an existing object.

fn delete_objects(
&self,
input: &DeleteObjectsRequest
) -> Result<DeleteObjectsOutput, DeleteObjectsError>

This operation enables you to delete multiple objects from a bucket using a single HTTP request. You may specify up to 1000 keys.

fn get_bucket_accelerate_configuration(
&self,
input: &GetBucketAccelerateConfigurationRequest
) -> Result<GetBucketAccelerateConfigurationOutput, GetBucketAccelerateConfigurationError>

Returns the accelerate configuration of a bucket.

fn get_bucket_acl(
&self,
input: &GetBucketAclRequest
) -> Result<GetBucketAclOutput, GetBucketAclError>

Gets the access control policy for the bucket.

fn get_bucket_analytics_configuration(
&self,
input: &GetBucketAnalyticsConfigurationRequest
) -> Result<GetBucketAnalyticsConfigurationOutput, GetBucketAnalyticsConfigurationError>

Gets an analytics configuration for the bucket (specified by the analytics configuration ID).

fn get_bucket_cors(
&self,
input: &GetBucketCorsRequest
) -> Result<GetBucketCorsOutput, GetBucketCorsError>

Returns the cors configuration for the bucket.

fn get_bucket_inventory_configuration(
&self,
input: &GetBucketInventoryConfigurationRequest
) -> Result<GetBucketInventoryConfigurationOutput, GetBucketInventoryConfigurationError>

Returns an inventory configuration (identified by the inventory ID) from the bucket.

fn get_bucket_lifecycle(
&self,
input: &GetBucketLifecycleRequest
) -> Result<GetBucketLifecycleOutput, GetBucketLifecycleError>

Deprecated, see the GetBucketLifecycleConfiguration operation.

fn get_bucket_lifecycle_configuration(
&self,
input: &GetBucketLifecycleConfigurationRequest
) -> Result<GetBucketLifecycleConfigurationOutput, GetBucketLifecycleConfigurationError>

Returns the lifecycle configuration information set on the bucket.

fn get_bucket_location(
&self,
input: &GetBucketLocationRequest
) -> Result<GetBucketLocationOutput, GetBucketLocationError>

Returns the region the bucket resides in.

fn get_bucket_logging(
&self,
input: &GetBucketLoggingRequest
) -> Result<GetBucketLoggingOutput, GetBucketLoggingError>

Returns the logging status of a bucket and the permissions users have to view and modify that status. To use GET, you must be the bucket owner.

fn get_bucket_metrics_configuration(
&self,
input: &GetBucketMetricsConfigurationRequest
) -> Result<GetBucketMetricsConfigurationOutput, GetBucketMetricsConfigurationError>

Gets a metrics configuration (specified by the metrics configuration ID) from the bucket.

fn get_bucket_notification(
&self,
input: &GetBucketNotificationConfigurationRequest
) -> Result<NotificationConfigurationDeprecated, GetBucketNotificationError>

Deprecated, see the GetBucketNotificationConfiguration operation.

fn get_bucket_notification_configuration(
&self,
input: &GetBucketNotificationConfigurationRequest
) -> Result<NotificationConfiguration, GetBucketNotificationConfigurationError>

Returns the notification configuration of a bucket.

fn get_bucket_policy(
&self,
input: &GetBucketPolicyRequest
) -> Result<GetBucketPolicyOutput, GetBucketPolicyError>

Returns the policy of a specified bucket.

fn get_bucket_replication(
&self,
input: &GetBucketReplicationRequest
) -> Result<GetBucketReplicationOutput, GetBucketReplicationError>

Returns the replication configuration of a bucket.

fn get_bucket_request_payment(
&self,
input: &GetBucketRequestPaymentRequest
) -> Result<GetBucketRequestPaymentOutput, GetBucketRequestPaymentError>

Returns the request payment configuration of a bucket.

fn get_bucket_tagging(
&self,
input: &GetBucketTaggingRequest
) -> Result<GetBucketTaggingOutput, GetBucketTaggingError>

Returns the tag set associated with the bucket.

fn get_bucket_versioning(
&self,
input: &GetBucketVersioningRequest
) -> Result<GetBucketVersioningOutput, GetBucketVersioningError>

Returns the versioning state of a bucket.

fn get_bucket_website(
&self,
input: &GetBucketWebsiteRequest
) -> Result<GetBucketWebsiteOutput, GetBucketWebsiteError>

Returns the website configuration for a bucket.

fn get_object(
&self,
input: &GetObjectRequest
) -> Result<GetObjectOutput, GetObjectError>

Retrieves objects from Amazon S3.

fn get_object_acl(
&self,
input: &GetObjectAclRequest
) -> Result<GetObjectAclOutput, GetObjectAclError>

Returns the access control list (ACL) of an object.

fn get_object_tagging(
&self,
input: &GetObjectTaggingRequest
) -> Result<GetObjectTaggingOutput, GetObjectTaggingError>

Returns the tag-set of an object.

fn get_object_torrent(
&self,
input: &GetObjectTorrentRequest
) -> Result<GetObjectTorrentOutput, GetObjectTorrentError>

Return torrent files from a bucket.

fn head_bucket(&self, input: &HeadBucketRequest) -> Result<(), HeadBucketError>

This operation is useful to determine if a bucket exists and you have permission to access it.

fn head_object(
&self,
input: &HeadObjectRequest
) -> Result<HeadObjectOutput, HeadObjectError>

The HEAD operation retrieves metadata from an object without returning the object itself. This operation is useful if you're only interested in an object's metadata. To use HEAD, you must have READ access to the object.

fn list_bucket_analytics_configurations(
&self,
input: &ListBucketAnalyticsConfigurationsRequest
) -> Result<ListBucketAnalyticsConfigurationsOutput, ListBucketAnalyticsConfigurationsError>

Lists the analytics configurations for the bucket.

fn list_bucket_inventory_configurations(
&self,
input: &ListBucketInventoryConfigurationsRequest
) -> Result<ListBucketInventoryConfigurationsOutput, ListBucketInventoryConfigurationsError>

Returns a list of inventory configurations for the bucket.

fn list_bucket_metrics_configurations(
&self,
input: &ListBucketMetricsConfigurationsRequest
) -> Result<ListBucketMetricsConfigurationsOutput, ListBucketMetricsConfigurationsError>

Lists the metrics configurations for the bucket.

fn list_buckets(&self) -> Result<ListBucketsOutput, ListBucketsError>

Returns a list of all buckets owned by the authenticated sender of the request.

fn list_multipart_uploads(
&self,
input: &ListMultipartUploadsRequest
) -> Result<ListMultipartUploadsOutput, ListMultipartUploadsError>

This operation lists in-progress multipart uploads.

fn list_object_versions(
&self,
input: &ListObjectVersionsRequest
) -> Result<ListObjectVersionsOutput, ListObjectVersionsError>

Returns metadata about all of the versions of objects in a bucket.

fn list_objects(
&self,
input: &ListObjectsRequest
) -> Result<ListObjectsOutput, ListObjectsError>

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.

fn list_objects_v2(
&self,
input: &ListObjectsV2Request
) -> Result<ListObjectsV2Output, ListObjectsV2Error>

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket. Note: ListObjectsV2 is the revised List Objects API and we recommend you use this revised API for new application development.

fn list_parts(
&self,
input: &ListPartsRequest
) -> Result<ListPartsOutput, ListPartsError>

Lists the parts that have been uploaded for a specific multipart upload.

fn put_bucket_accelerate_configuration(
&self,
input: &PutBucketAccelerateConfigurationRequest
) -> Result<(), PutBucketAccelerateConfigurationError>

Sets the accelerate configuration of an existing bucket.

fn put_bucket_acl(
&self,
input: &PutBucketAclRequest
) -> Result<(), PutBucketAclError>

Sets the permissions on a bucket using access control lists (ACL).

fn put_bucket_analytics_configuration(
&self,
input: &PutBucketAnalyticsConfigurationRequest
) -> Result<(), PutBucketAnalyticsConfigurationError>

Sets an analytics configuration for the bucket (specified by the analytics configuration ID).

fn put_bucket_cors(
&self,
input: &PutBucketCorsRequest
) -> Result<(), PutBucketCorsError>

Sets the cors configuration for a bucket.

fn put_bucket_inventory_configuration(
&self,
input: &PutBucketInventoryConfigurationRequest
) -> Result<(), PutBucketInventoryConfigurationError>

Adds an inventory configuration (identified by the inventory ID) from the bucket.

fn put_bucket_lifecycle(
&self,
input: &PutBucketLifecycleRequest
) -> Result<(), PutBucketLifecycleError>

Deprecated, see the PutBucketLifecycleConfiguration operation.

fn put_bucket_lifecycle_configuration(
&self,
input: &PutBucketLifecycleConfigurationRequest
) -> Result<(), PutBucketLifecycleConfigurationError>

Sets lifecycle configuration for your bucket. If a lifecycle configuration exists, it replaces it.

fn put_bucket_logging(
&self,
input: &PutBucketLoggingRequest
) -> Result<(), PutBucketLoggingError>

Set the logging parameters for a bucket and to specify permissions for who can view and modify the logging parameters. To set the logging status of a bucket, you must be the bucket owner.

fn put_bucket_metrics_configuration(
&self,
input: &PutBucketMetricsConfigurationRequest
) -> Result<(), PutBucketMetricsConfigurationError>

Sets a metrics configuration (specified by the metrics configuration ID) for the bucket.

fn put_bucket_notification(
&self,
input: &PutBucketNotificationRequest
) -> Result<(), PutBucketNotificationError>

Deprecated, see the PutBucketNotificationConfiguraiton operation.

fn put_bucket_notification_configuration(
&self,
input: &PutBucketNotificationConfigurationRequest
) -> Result<(), PutBucketNotificationConfigurationError>

Enables notifications of specified events for a bucket.

fn put_bucket_policy(
&self,
input: &PutBucketPolicyRequest
) -> Result<(), PutBucketPolicyError>

Replaces a policy on a bucket. If the bucket already has a policy, the one in this request completely replaces it.

fn put_bucket_replication(
&self,
input: &PutBucketReplicationRequest
) -> Result<(), PutBucketReplicationError>

Creates a new replication configuration (or replaces an existing one, if present).

fn put_bucket_request_payment(
&self,
input: &PutBucketRequestPaymentRequest
) -> Result<(), PutBucketRequestPaymentError>

Sets the request payment configuration for a bucket. By default, the bucket owner pays for downloads from the bucket. This configuration parameter enables the bucket owner (only) to specify that the person requesting the download will be charged for the download. Documentation on requester pays buckets can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/RequesterPaysBuckets.html

fn put_bucket_tagging(
&self,
input: &PutBucketTaggingRequest
) -> Result<(), PutBucketTaggingError>

Sets the tags for a bucket.

fn put_bucket_versioning(
&self,
input: &PutBucketVersioningRequest
) -> Result<(), PutBucketVersioningError>

Sets the versioning state of an existing bucket. To set the versioning state, you must be the bucket owner.

fn put_bucket_website(
&self,
input: &PutBucketWebsiteRequest
) -> Result<(), PutBucketWebsiteError>

Set the website configuration for a bucket.

fn put_object(
&self,
input: &PutObjectRequest
) -> Result<PutObjectOutput, PutObjectError>

Adds an object to a bucket.

fn put_object_acl(
&self,
input: &PutObjectAclRequest
) -> Result<PutObjectAclOutput, PutObjectAclError>

uses the acl subresource to set the access control list (ACL) permissions for an object that already exists in a bucket

fn put_object_tagging(
&self,
input: &PutObjectTaggingRequest
) -> Result<PutObjectTaggingOutput, PutObjectTaggingError>

Sets the supplied tag-set to an object that already exists in a bucket

fn restore_object(
&self,
input: &RestoreObjectRequest
) -> Result<RestoreObjectOutput, RestoreObjectError>

Restores an archived copy of an object back into Amazon S3

fn upload_part(
&self,
input: &UploadPartRequest
) -> Result<UploadPartOutput, UploadPartError>

Uploads a part in a multipart upload.

Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

fn upload_part_copy(
&self,
input: &UploadPartCopyRequest
) -> Result<UploadPartCopyOutput, UploadPartCopyError>

Uploads a part by copying data from an existing object as data source.