Crate rusoto_s3

If you're using the service, you're probably looking for S3Client and S3.

Modules

util
Utility helpers for working with S3

Structs

AbortIncompleteMultipartUpload
Specifies the days since the initiation of an incomplete multipart upload that Amazon S3 will wait before permanently removing all parts of the upload. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy in the Amazon Simple Storage Service Developer Guide.

AbortIncompleteMultipartUploadSerializer

AbortMultipartUploadOutput

AbortMultipartUploadRequest

AccelerateConfiguration
Configures the transfer acceleration state for an Amazon S3 bucket. For more information, see Amazon S3 Transfer Acceleration in the Amazon Simple Storage Service Developer Guide.

AccelerateConfigurationSerializer

AccessControlPolicy
Contains the elements that set the ACL permissions for an object per grantee.

AccessControlPolicySerializer

AccessControlTranslation
A container for information about access control for replicas.

AccessControlTranslationSerializer

AccountIdSerializer

AllowQuotedRecordDelimiterSerializer

AllowedHeaderSerializer

AllowedHeadersSerializer

AllowedMethodSerializer

AllowedMethodsSerializer

AllowedOriginSerializer

AllowedOriginsSerializer

AnalyticsAndOperator
A conjunction (logical AND) of predicates, which is used in evaluating a metrics filter. The operator must have at least two predicates in any combination, and an object must match all of the predicates for the filter to apply.

AnalyticsAndOperatorSerializer

AnalyticsConfiguration
Specifies the configuration and any analyses for the analytics filter of an Amazon S3 bucket.

For more information, see GET Bucket analytics in the Amazon Simple Storage Service API Reference.

AnalyticsConfigurationSerializer

AnalyticsExportDestination
Where to publish the analytics results.

AnalyticsExportDestinationSerializer

AnalyticsFilter

AnalyticsFilterSerializer

AnalyticsIdSerializer

AnalyticsS3BucketDestination

AnalyticsS3BucketDestinationSerializer

AnalyticsS3ExportFileFormatSerializer

BodySerializer

Bucket

BucketAccelerateStatusSerializer

BucketLifecycleConfiguration
Specifies the lifecycle configuration for objects in an Amazon S3 bucket. For more information, see Object Lifecycle Management in the Amazon Simple Storage Service Developer Guide.

BucketLifecycleConfigurationSerializer

BucketLocationConstraintSerializer

BucketLoggingStatus

BucketLoggingStatusSerializer

BucketLogsPermissionSerializer

BucketNameSerializer

BucketVersioningStatusSerializer

CORSConfiguration
Describes the cross-origin access configuration for objects in an Amazon S3 bucket. For more information, see Enabling Cross-Origin Resource Sharing in the Amazon Simple Storage Service Developer Guide.

CORSConfigurationSerializer

CORSRule
Specifies a cross-origin access rule for an Amazon S3 bucket.

CORSRuleSerializer

CORSRulesSerializer

CSVInput
Describes how a CSV-formatted input object is formatted.

CSVInputSerializer

CSVOutput
Describes how CSV-formatted results are formatted.

CSVOutputSerializer

CloudFunctionConfiguration

CloudFunctionConfigurationSerializer

CloudFunctionInvocationRoleSerializer

CloudFunctionSerializer

CommentsSerializer

CommonPrefix

CompleteMultipartUploadOutput

CompleteMultipartUploadRequest

CompletedMultipartUpload

CompletedMultipartUploadSerializer

CompletedPart

CompletedPartListSerializer

CompletedPartSerializer

CompressionTypeSerializer

Condition
Specifies a condition that must be met for a redirect to apply.

ConditionSerializer

ContinuationEvent

CopyObjectOutput

CopyObjectRequest

CopyObjectResult

CopyPartResult

CreateBucketConfiguration

CreateBucketConfigurationSerializer

CreateBucketOutput

CreateBucketRequest

CreateMultipartUploadOutput

CreateMultipartUploadRequest

DateSerializer

DaysAfterInitiationSerializer

DaysSerializer

DefaultRetention
The container element for specifying the default object lock retention settings for new objects placed in the specified bucket.

DefaultRetentionSerializer

Delete

DeleteBucketAnalyticsConfigurationRequest

DeleteBucketCorsRequest

DeleteBucketEncryptionRequest

DeleteBucketInventoryConfigurationRequest

DeleteBucketLifecycleRequest

DeleteBucketMetricsConfigurationRequest

DeleteBucketPolicyRequest

DeleteBucketReplicationRequest

DeleteBucketRequest

DeleteBucketTaggingRequest

DeleteBucketWebsiteRequest

DeleteMarkerEntry

DeleteMarkerReplication
Specifies whether Amazon S3 should replicate delete makers.

DeleteMarkerReplicationSerializer

DeleteMarkerReplicationStatusSerializer

DeleteObjectOutput

DeleteObjectRequest

DeleteObjectTaggingOutput

DeleteObjectTaggingRequest

DeleteObjectsOutput

DeleteObjectsRequest

DeletePublicAccessBlockRequest

DeleteSerializer

DeletedObject

DelimiterSerializer

DescriptionSerializer

Destination
Specifies information about where to publish analysis or configuration results for an Amazon S3 bucket.

DestinationSerializer

DisplayNameSerializer

ETagSerializer

EmailAddressSerializer

EnableRequestProgressSerializer

EncodingTypeSerializer

Encryption
Describes the server-side encryption that will be applied to the restore results.

EncryptionConfiguration
Specifies encryption-related information for an Amazon S3 bucket that is a destination for replicated objects.

EncryptionConfigurationSerializer

EncryptionSerializer

EndEvent

ErrorDocument

ErrorDocumentSerializer

EventListSerializer

EventSerializer

ExpirationStatusSerializer

ExpiredObjectDeleteMarkerSerializer

ExposeHeaderSerializer

ExposeHeadersSerializer

ExpressionSerializer

ExpressionTypeSerializer

FetchOwnerSerializer

FieldDelimiterSerializer

FileHeaderInfoSerializer

FilterRule
Specifies the Amazon S3 object key name to filter on and whether to filter on the suffix or prefix of the key name.

FilterRuleListSerializer

FilterRuleNameSerializer

FilterRuleSerializer

FilterRuleValueSerializer

GetBucketAccelerateConfigurationOutput

GetBucketAccelerateConfigurationRequest

GetBucketAclOutput

GetBucketAclRequest

GetBucketAnalyticsConfigurationOutput

GetBucketAnalyticsConfigurationRequest

GetBucketCorsOutput

GetBucketCorsRequest

GetBucketEncryptionOutput

GetBucketEncryptionRequest

GetBucketInventoryConfigurationOutput

GetBucketInventoryConfigurationRequest

GetBucketLifecycleConfigurationOutput

GetBucketLifecycleConfigurationRequest

GetBucketLifecycleOutput

GetBucketLifecycleRequest

GetBucketLocationOutput

GetBucketLocationRequest

GetBucketLoggingOutput

GetBucketLoggingRequest

GetBucketMetricsConfigurationOutput

GetBucketMetricsConfigurationRequest

GetBucketNotificationConfigurationRequest

GetBucketPolicyOutput

GetBucketPolicyRequest

GetBucketPolicyStatusOutput

GetBucketPolicyStatusRequest

GetBucketReplicationOutput

GetBucketReplicationRequest

GetBucketRequestPaymentOutput

GetBucketRequestPaymentRequest

GetBucketTaggingOutput

GetBucketTaggingRequest

GetBucketVersioningOutput

GetBucketVersioningRequest

GetBucketWebsiteOutput

GetBucketWebsiteRequest

GetObjectAclOutput

GetObjectAclRequest

GetObjectLegalHoldOutput

GetObjectLegalHoldRequest

GetObjectLockConfigurationOutput

GetObjectLockConfigurationRequest

GetObjectOutput

GetObjectRequest

GetObjectRetentionOutput

GetObjectRetentionRequest

GetObjectTaggingOutput

GetObjectTaggingRequest

GetObjectTorrentOutput

GetObjectTorrentRequest

GetPublicAccessBlockOutput

GetPublicAccessBlockRequest

GlacierJobParameters

GlacierJobParametersSerializer

Grant

GrantSerializer

Grantee

GranteeSerializer

GrantsSerializer

HeadBucketRequest

HeadObjectOutput

HeadObjectRequest

HostNameSerializer

HttpErrorCodeReturnedEqualsSerializer

HttpRedirectCodeSerializer

IDSerializer

IndexDocument

IndexDocumentSerializer

Initiator

InputSerialization
Describes the serialization format of the object.

InputSerializationSerializer

InventoryConfiguration
Specifies the inventory configuration for an Amazon S3 bucket. For more information, see GET Bucket inventory in the Amazon Simple Storage Service API Reference.

InventoryConfigurationSerializer

InventoryDestination

InventoryDestinationSerializer

InventoryEncryption
Contains the type of server-side encryption used to encrypt the inventory results.

InventoryEncryptionSerializer

InventoryFilter

InventoryFilterSerializer

InventoryFormatSerializer

InventoryFrequencySerializer

InventoryIdSerializer

InventoryIncludedObjectVersionsSerializer

InventoryOptionalFieldSerializer

InventoryOptionalFieldsSerializer

InventoryS3BucketDestination

InventoryS3BucketDestinationSerializer

InventorySchedule

InventoryScheduleSerializer

IsEnabledSerializer

JSONInput

JSONInputSerializer

JSONOutput

JSONOutputSerializer

JSONTypeSerializer

KMSContextSerializer

KeyMarkerSerializer

KeyPrefixEqualsSerializer

LambdaFunctionArnSerializer

LambdaFunctionConfiguration
A container for specifying the configuration for AWS Lambda notifications.

LambdaFunctionConfigurationListSerializer

LambdaFunctionConfigurationSerializer

LifecycleConfiguration

LifecycleConfigurationSerializer

LifecycleExpiration

LifecycleExpirationSerializer

LifecycleRule

LifecycleRuleAndOperator
This is used in a Lifecycle Rule Filter to apply a logical AND to two or more predicates. The Lifecycle Rule will apply to any object matching all of the predicates configured inside the And operator.

LifecycleRuleAndOperatorSerializer

LifecycleRuleFilter
The Filter is used to identify objects that a Lifecycle Rule applies to. A Filter must have exactly one of Prefix, Tag, or And specified.

LifecycleRuleFilterSerializer

LifecycleRuleSerializer

LifecycleRulesSerializer

ListBucketAnalyticsConfigurationsOutput

ListBucketAnalyticsConfigurationsRequest

ListBucketInventoryConfigurationsOutput

ListBucketInventoryConfigurationsRequest

ListBucketMetricsConfigurationsOutput

ListBucketMetricsConfigurationsRequest

ListBucketsOutput

ListMultipartUploadsOutput

ListMultipartUploadsRequest

ListObjectVersionsOutput

ListObjectVersionsRequest

ListObjectsOutput

ListObjectsRequest

ListObjectsV2Output

ListObjectsV2Request

ListPartsOutput

ListPartsRequest

LocationPrefixSerializer

LoggingEnabled
Describes where logs are stored and the prefix that Amazon S3 assigns to all log object keys for a bucket. For more information, see PUT Bucket logging in the Amazon Simple Storage Service API Reference.

LoggingEnabledSerializer

MFADeleteSerializer

MarkerSerializer

MaxAgeSecondsSerializer

MaxKeysSerializer

MaxPartsSerializer

MaxUploadsSerializer

MetadataEntry
A metadata key-value pair to store with an object.

MetadataEntrySerializer

MetadataKeySerializer

MetadataValueSerializer

MetricsAndOperator

MetricsAndOperatorSerializer

MetricsConfiguration
Specifies a metrics configuration for the CloudWatch request metrics (specified by the metrics configuration ID) from an Amazon S3 bucket. If you're updating an existing metrics configuration, note that this is a full replacement of the existing metrics configuration. If you don't include the elements you want to keep, they are erased. For more information, see PUT Bucket metrics in the Amazon Simple Storage Service API Reference.

MetricsConfigurationSerializer

MetricsFilter

MetricsFilterSerializer

MetricsIdSerializer

MultipartUpload

MultipartUploadIdSerializer

NoncurrentVersionExpiration
Specifies when noncurrent object versions expire. Upon expiration, Amazon S3 permanently deletes the noncurrent object versions. You set this lifecycle configuration action on a bucket that has versioning enabled (or suspended) to request that Amazon S3 delete noncurrent object versions at a specific period in the object's lifetime.

NoncurrentVersionExpirationSerializer

NoncurrentVersionTransition
Container for the transition rule that describes when noncurrent objects transition to the STANDARD_IA, ONEZONE_IA, INTELLIGENT_TIERING, GLACIER, or DEEP_ARCHIVE storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to the STANDARD_IA, ONEZONE_IA, INTELLIGENT_TIERING, GLACIER, or DEEP_ARCHIVE storage class at a specific period in the object's lifetime.

NoncurrentVersionTransitionListSerializer

NoncurrentVersionTransitionSerializer

NotificationConfiguration
A container for specifying the notification configuration of the bucket. If this element is empty, notifications are turned off for the bucket.

NotificationConfigurationDeprecated

NotificationConfigurationDeprecatedSerializer

NotificationConfigurationFilter
Specifies object key name filtering rules. For information about key name filtering, see Configuring Event Notifications in the Amazon Simple Storage Service Developer Guide.

NotificationConfigurationFilterSerializer

NotificationConfigurationSerializer

NotificationIdSerializer

Object

ObjectCannedACLSerializer

ObjectIdentifier

ObjectIdentifierListSerializer

ObjectIdentifierSerializer

ObjectKeySerializer

ObjectLockConfiguration
The container element for object lock configuration parameters.

ObjectLockConfigurationSerializer

ObjectLockEnabledSerializer

ObjectLockLegalHold
A Legal Hold configuration for an object.

ObjectLockLegalHoldSerializer

ObjectLockLegalHoldStatusSerializer

ObjectLockRetention
A Retention configuration for an object.

ObjectLockRetentionModeSerializer

ObjectLockRetentionSerializer

ObjectLockRule
The container element for an object lock rule.

ObjectLockRuleSerializer

ObjectVersion

ObjectVersionIdSerializer

OutputLocation
Describes the location where the restore job's output is stored.

OutputLocationSerializer

OutputSerialization
Describes how results of the Select job are serialized.

OutputSerializationSerializer

Owner

OwnerOverrideSerializer

OwnerSerializer

ParquetInput

ParquetInputSerializer

Part

PartNumberMarkerSerializer

PartNumberSerializer

PayerSerializer

PermissionSerializer

PolicyStatus
The container element for a bucket's policy status.

PrefixSerializer

PrioritySerializer

Progress

ProgressEvent

ProtocolSerializer

PublicAccessBlockConfiguration
Specifies the Block Public Access configuration for an Amazon S3 bucket.

PublicAccessBlockConfigurationSerializer

PutBucketAccelerateConfigurationRequest

PutBucketAclRequest

PutBucketAnalyticsConfigurationRequest

PutBucketCorsRequest

PutBucketEncryptionRequest

PutBucketInventoryConfigurationRequest

PutBucketLifecycleConfigurationRequest

PutBucketLifecycleRequest

PutBucketLoggingRequest

PutBucketMetricsConfigurationRequest

PutBucketNotificationConfigurationRequest

PutBucketNotificationRequest

PutBucketPolicyRequest

PutBucketReplicationRequest

PutBucketRequestPaymentRequest

PutBucketTaggingRequest

PutBucketVersioningRequest

PutBucketWebsiteRequest

PutObjectAclOutput

PutObjectAclRequest

PutObjectLegalHoldOutput

PutObjectLegalHoldRequest

PutObjectLockConfigurationOutput

PutObjectLockConfigurationRequest

PutObjectOutput

PutObjectRequest

PutObjectRetentionOutput

PutObjectRetentionRequest

PutObjectTaggingOutput

PutObjectTaggingRequest

PutPublicAccessBlockRequest

QueueArnSerializer

QueueConfiguration
Specifies the configuration for publishing messages to an Amazon Simple Queue Service (Amazon SQS) queue when Amazon S3 detects specified events.

QueueConfigurationDeprecated

QueueConfigurationDeprecatedSerializer

QueueConfigurationListSerializer

QueueConfigurationSerializer

QuietSerializer

QuoteCharacterSerializer

QuoteEscapeCharacterSerializer

QuoteFieldsSerializer

RecordDelimiterSerializer

RecordsEvent

Redirect
Specifies how requests are redirected. In the event of an error, you can specify a different error code to return.

RedirectAllRequestsTo
Specifies the redirect behavior of all requests to a website endpoint of an Amazon S3 bucket.

RedirectAllRequestsToSerializer

RedirectSerializer

ReplaceKeyPrefixWithSerializer

ReplaceKeyWithSerializer

ReplicaKmsKeyIDSerializer

ReplicationConfiguration
A container for replication rules. You can add up to 1,000 rules. The maximum size of a replication configuration is 2 MB.

ReplicationConfigurationSerializer

ReplicationRule
Specifies which Amazon S3 objects to replicate and where to store the replicas.

ReplicationRuleAndOperator

ReplicationRuleAndOperatorSerializer

ReplicationRuleFilter
A filter that identifies the subset of objects to which the replication rule applies. A Filter must specify exactly one Prefix, Tag, or an And child element.

ReplicationRuleFilterSerializer

ReplicationRuleSerializer

ReplicationRuleStatusSerializer

ReplicationRulesSerializer

RequestPaymentConfiguration

RequestPaymentConfigurationSerializer

RequestProgress

RequestProgressSerializer

ResponseCacheControlSerializer

ResponseContentDispositionSerializer

ResponseContentEncodingSerializer

ResponseContentLanguageSerializer

ResponseContentTypeSerializer

ResponseExpiresSerializer

RestoreObjectOutput

RestoreObjectRequest

RestoreRequest
Container for restore job parameters.

RestoreRequestSerializer

RestoreRequestTypeSerializer

RoleSerializer

RoutingRule
Specifies the redirect behavior and when a redirect is applied.

RoutingRuleSerializer

RoutingRulesSerializer

Rule
Specifies lifecycle rules for an Amazon S3 bucket. For more information, see PUT Bucket lifecycle in the Amazon Simple Storage Service API Reference.

RuleSerializer

RulesSerializer

S3Error

S3KeyFilter
A container for object key name prefix and suffix filtering rules.

S3KeyFilterSerializer

S3Location
Describes an S3 location that will receive the results of the restore request.

S3LocationSerializer

S3Client
A client for the Amazon S3 API.

SSEKMS
Specifies the use of SSE-KMS to encrypt delivered Inventory reports.

SSEKMSKeyIdSerializer

SSEKMSSerializer

SSES3
Specifies the use of SSE-S3 to encrypt delivered Inventory reports.

SSES3Serializer

SelectObjectContentEventStream

SelectObjectContentOutput

SelectObjectContentRequest
Request to filter the contents of an Amazon S3 object based on a simple Structured Query Language (SQL) statement. In the request, along with the SQL expression, you must specify a data serialization format (JSON or CSV) of the object. Amazon S3 uses this to parse object data into records. It returns only records that match the specified SQL expression. You must also specify the data serialization format for the response. For more information, see S3Select API Documentation.

SelectObjectContentRequestSerializer

SelectParameters
Describes the parameters for Select job types.

SelectParametersSerializer

ServerSideEncryptionByDefault
Describes the default server-side encryption to apply to new objects in the bucket. If a PUT Object request doesn't specify any server-side encryption, this default encryption will be applied. For more information, see PUT Bucket encryption in the Amazon Simple Storage Service API Reference.

ServerSideEncryptionByDefaultSerializer

ServerSideEncryptionConfiguration
Specifies the default server-side-encryption configuration.

ServerSideEncryptionConfigurationSerializer

ServerSideEncryptionRule
Specifies the default server-side encryption configuration.

ServerSideEncryptionRuleSerializer

ServerSideEncryptionRulesSerializer

ServerSideEncryptionSerializer

SettingSerializer

SourceSelectionCriteria
A container that describes additional filters for identifying the source objects that you want to replicate. You can choose to enable or disable the replication of these objects. Currently, Amazon S3 supports only the filter that you can specify for objects created with server-side encryption using an AWS KMS-Managed Key (SSE-KMS).

SourceSelectionCriteriaSerializer

SseKmsEncryptedObjects
A container for filter information for the selection of S3 objects encrypted with AWS KMS.

SseKmsEncryptedObjectsSerializer

SseKmsEncryptedObjectsStatusSerializer

StartAfterSerializer

Stats

StatsEvent

StorageClassAnalysis
Specifies data related to access patterns to be collected and made available to analyze the tradeoffs between different storage classes for an Amazon S3 bucket.

StorageClassAnalysisDataExport

StorageClassAnalysisDataExportSerializer

StorageClassAnalysisSchemaVersionSerializer

StorageClassAnalysisSerializer

StorageClassSerializer

SuffixSerializer

Tag

TagSerializer

TagSetSerializer

Tagging

TaggingSerializer

TargetBucketSerializer

TargetGrant

TargetGrantSerializer

TargetGrantsSerializer

TargetPrefixSerializer

TierSerializer

TokenSerializer

TopicArnSerializer

TopicConfiguration
A container for specifying the configuration for publication of messages to an Amazon Simple Notification Service (Amazon SNS) topic when Amazon S3 detects specified events.

TopicConfigurationDeprecated

TopicConfigurationDeprecatedSerializer

TopicConfigurationListSerializer

TopicConfigurationSerializer

Transition
Specifies when an object transitions to a specified storage class.

TransitionListSerializer

TransitionSerializer

TransitionStorageClassSerializer

TypeSerializer

URISerializer

UploadIdMarkerSerializer

UploadPartCopyOutput

UploadPartCopyRequest

UploadPartOutput

UploadPartRequest

UserMetadataSerializer

ValueSerializer

VersionIdMarkerSerializer

VersioningConfiguration
Describes the versioning state of an Amazon S3 bucket. For more information, see PUT Bucket versioning in the Amazon Simple Storage Service API Reference.

VersioningConfigurationSerializer

WebsiteConfiguration
Specifies website configuration parameters for an Amazon S3 bucket.

WebsiteConfigurationSerializer

YearsSerializer

Enums

AbortMultipartUploadError
Errors returned by AbortMultipartUpload

CompleteMultipartUploadError
Errors returned by CompleteMultipartUpload

CopyObjectError
Errors returned by CopyObject

CreateBucketError
Errors returned by CreateBucket

CreateMultipartUploadError
Errors returned by CreateMultipartUpload

DeleteBucketAnalyticsConfigurationError
Errors returned by DeleteBucketAnalyticsConfiguration

DeleteBucketCorsError
Errors returned by DeleteBucketCors

DeleteBucketEncryptionError
Errors returned by DeleteBucketEncryption

DeleteBucketError
Errors returned by DeleteBucket

DeleteBucketInventoryConfigurationError
Errors returned by DeleteBucketInventoryConfiguration

DeleteBucketLifecycleError
Errors returned by DeleteBucketLifecycle

DeleteBucketMetricsConfigurationError
Errors returned by DeleteBucketMetricsConfiguration

DeleteBucketPolicyError
Errors returned by DeleteBucketPolicy

DeleteBucketReplicationError
Errors returned by DeleteBucketReplication

DeleteBucketTaggingError
Errors returned by DeleteBucketTagging

DeleteBucketWebsiteError
Errors returned by DeleteBucketWebsite

DeleteObjectError
Errors returned by DeleteObject

DeleteObjectTaggingError
Errors returned by DeleteObjectTagging

DeleteObjectsError
Errors returned by DeleteObjects

DeletePublicAccessBlockError
Errors returned by DeletePublicAccessBlock

GetBucketAccelerateConfigurationError
Errors returned by GetBucketAccelerateConfiguration

GetBucketAclError
Errors returned by GetBucketAcl

GetBucketAnalyticsConfigurationError
Errors returned by GetBucketAnalyticsConfiguration

GetBucketCorsError
Errors returned by GetBucketCors

GetBucketEncryptionError
Errors returned by GetBucketEncryption

GetBucketInventoryConfigurationError
Errors returned by GetBucketInventoryConfiguration

GetBucketLifecycleConfigurationError
Errors returned by GetBucketLifecycleConfiguration

GetBucketLifecycleError
Errors returned by GetBucketLifecycle

GetBucketLocationError
Errors returned by GetBucketLocation

GetBucketLoggingError
Errors returned by GetBucketLogging

GetBucketMetricsConfigurationError
Errors returned by GetBucketMetricsConfiguration

GetBucketNotificationConfigurationError
Errors returned by GetBucketNotificationConfiguration

GetBucketNotificationError
Errors returned by GetBucketNotification

GetBucketPolicyError
Errors returned by GetBucketPolicy

GetBucketPolicyStatusError
Errors returned by GetBucketPolicyStatus

GetBucketReplicationError
Errors returned by GetBucketReplication

GetBucketRequestPaymentError
Errors returned by GetBucketRequestPayment

GetBucketTaggingError
Errors returned by GetBucketTagging

GetBucketVersioningError
Errors returned by GetBucketVersioning

GetBucketWebsiteError
Errors returned by GetBucketWebsite

GetObjectAclError
Errors returned by GetObjectAcl

GetObjectError
Errors returned by GetObject

GetObjectLegalHoldError
Errors returned by GetObjectLegalHold

GetObjectLockConfigurationError
Errors returned by GetObjectLockConfiguration

GetObjectRetentionError
Errors returned by GetObjectRetention

GetObjectTaggingError
Errors returned by GetObjectTagging

GetObjectTorrentError
Errors returned by GetObjectTorrent

GetPublicAccessBlockError
Errors returned by GetPublicAccessBlock

HeadBucketError
Errors returned by HeadBucket

HeadObjectError
Errors returned by HeadObject

ListBucketAnalyticsConfigurationsError
Errors returned by ListBucketAnalyticsConfigurations

ListBucketInventoryConfigurationsError
Errors returned by ListBucketInventoryConfigurations

ListBucketMetricsConfigurationsError
Errors returned by ListBucketMetricsConfigurations

ListBucketsError
Errors returned by ListBuckets

ListMultipartUploadsError
Errors returned by ListMultipartUploads

ListObjectVersionsError
Errors returned by ListObjectVersions

ListObjectsError
Errors returned by ListObjects

ListObjectsV2Error
Errors returned by ListObjectsV2

ListPartsError
Errors returned by ListParts

PutBucketAccelerateConfigurationError
Errors returned by PutBucketAccelerateConfiguration

PutBucketAclError
Errors returned by PutBucketAcl

PutBucketAnalyticsConfigurationError
Errors returned by PutBucketAnalyticsConfiguration

PutBucketCorsError
Errors returned by PutBucketCors

PutBucketEncryptionError
Errors returned by PutBucketEncryption

PutBucketInventoryConfigurationError
Errors returned by PutBucketInventoryConfiguration

PutBucketLifecycleConfigurationError
Errors returned by PutBucketLifecycleConfiguration

PutBucketLifecycleError
Errors returned by PutBucketLifecycle

PutBucketLoggingError
Errors returned by PutBucketLogging

PutBucketMetricsConfigurationError
Errors returned by PutBucketMetricsConfiguration

PutBucketNotificationConfigurationError
Errors returned by PutBucketNotificationConfiguration

PutBucketNotificationError
Errors returned by PutBucketNotification

PutBucketPolicyError
Errors returned by PutBucketPolicy

PutBucketReplicationError
Errors returned by PutBucketReplication

PutBucketRequestPaymentError
Errors returned by PutBucketRequestPayment

PutBucketTaggingError
Errors returned by PutBucketTagging

PutBucketVersioningError
Errors returned by PutBucketVersioning

PutBucketWebsiteError
Errors returned by PutBucketWebsite

PutObjectAclError
Errors returned by PutObjectAcl

PutObjectError
Errors returned by PutObject

PutObjectLegalHoldError
Errors returned by PutObjectLegalHold

PutObjectLockConfigurationError
Errors returned by PutObjectLockConfiguration

PutObjectRetentionError
Errors returned by PutObjectRetention

PutObjectTaggingError
Errors returned by PutObjectTagging

PutPublicAccessBlockError
Errors returned by PutPublicAccessBlock

RestoreObjectError
Errors returned by RestoreObject

SelectObjectContentError
Errors returned by SelectObjectContent

UploadPartCopyError
Errors returned by UploadPartCopy

UploadPartError
Errors returned by UploadPart

Traits

S3
Trait representing the capabilities of the Amazon S3 API. Amazon S3 clients implement this trait.

Type Definitions

StreamingBody