Struct rusoto_s3::S3Client

pub struct S3Client { /* fields omitted */ }

A client for the Amazon S3 API.

Implementations

impl S3Client

pub fn new(region: Region) -> S3Client

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> S3Client where
P: ProvideAwsCredentials + Send + Sync + 'static,
D: DispatchSignedRequest + Send + Sync + 'static,

pub fn new_with_client(client: Client, region: Region) -> S3Client